Revert "First View Modification"
