+++ .env
@@ -1,8 +1,7 @@
 REACT_APP_NODE_ENV=production
 REACT_APP_HOST=127.0.0.1
 REACT_APP_PORT=3000
-# REACT_APP_URLBACKENDLANGUAGE="https://app.variamos.com/variamos_ms_languages"
-REACT_APP_URLBACKENDLANGUAGE="http://localhost:4000"
+REACT_APP_URLBACKENDLANGUAGE="https://app.variamos.com/variamos_ms_languages"
 REACT_APP_URLBACKENDRESTRICTION="https://app.variamos.com/vms_restrictions"
 REACT_APP_URLVARIAMOSLANGUAGES="https://app.variamos.com/variamos_languages"
 REACT_APP_URLVMSPROJECTS="https://app.variamos.com/vms_projects"
@@ -12,6 +11,6 @@ REACT_APP_URLVARIAMOSLANGDOCUMENTATION="https://www.variamos.com"
 
 REACT_APP_VARIAMOS_ADMIN_API_URL="https://app.variamos.com/variamos_ms_admin"
 REACT_APP_VARIAMOS_LOGIN_URL="https://app.variamos.com/variamos_admin/#/login"
-
+ 
 # Disables the login and allows you to log in as a "Local user"
 REACT_APP_DISABLE_LOGIN=false
