--- cypress.json
@@ -1,10 +1,0 @@
-{
-    "chromeWebSecurity": false,
-    "defaultCommandTimeout": 10000,
-    "pageLoadTimeout": 60000,
-    "e2e": {
-        "baseUrl": "http://localhost:3000",
-        "supportFile": "cypress/support/index.js",
-        "specPattern": "cypress/integration/**/*.js"
-    }
-}
