+++ src/App.tsx
@@ -1,8 +1,6 @@
 import {
   AnalyticsProvider,
   SessionProvider,
-  ResponseModel,
-  SessionInfoResponse,
 } from "@variamosple/variamos-components";
 import { FC } from "react";
 import { HashRouter, useRoutes } from "react-router-dom";
@@ -19,33 +17,12 @@ const Routes: FC = () => {
   return useRoutes(ROUTES);
 };
 
-// Check if running in Cypress test environment
-const isCypressTest = typeof window !== 'undefined' && (window as any).Cypress;
-
-// Mock getSessionInfo for Cypress tests
-const mockGetSessionInfo = async (): Promise<ResponseModel<SessionInfoResponse>> => {
-  const mockResponse = new ResponseModel<SessionInfoResponse>("SUCCESS");
-  mockResponse.data = {
-    user: {
-      id: 'guest-user',
-      name: 'Guest',
-      email: 'guest@example.com',
-      roles: ['guest'],
-      user: 'guest-user',
-      permissions: []
-    },
-    authenticated: true,
-    token: 'guest-token'
-  } as SessionInfoResponse;
-  return mockResponse;
-};
-
 const App: FC = () => {
   return (
     <AnalyticsProvider onVisit={registerVisit}>
       <SessionProvider
         loginUrl={AppConfig.LOGIN_URL}
-        getSessionInfo={isCypressTest ? mockGetSessionInfo : getSessionInfo}
+        getSessionInfo={getSessionInfo}
         requestLogout={requestLogout}
       >
         <HashRouter>
