+++ src/DataProvider/Services/languagesService.ts
@@ -3,7 +3,7 @@ import axios from "axios";
 import { ResponseModel } from "@variamosple/variamos-components";
 import { Language } from "../../Domain/ProductLineEngineering/Entities/Language";
 import { LANGUAGES_CLIENT } from "../../Infraestructure/AxiosConfig";
-import { LanguagesFilter } from "../../core/components/LanguageTable/LanguagesContainer";
+import { LanguagesFilter } from "../../core/components/PublicLanguages/PublicLanguagesContainer";
 
 export const queryUserLanguages = (
   filter: LanguagesFilter
@@ -45,39 +45,6 @@ export const queryLanguages = (
   filter: LanguagesFilter
 ): Promise<ResponseModel<Language[]>> => {
   return LANGUAGES_CLIENT.get(`/v2/languages`, {
-    params: filter,
-  })
-    .then((response) => response.data)
-    .catch((error) => {
-      if (axios.isAxiosError(error)) {
-        console.error("Axios error:", error.message);
-
-        const response = error.response?.data;
-
-        if (!!response) {
-          return response;
-        }
-
-        return new ResponseModel("BACK-ERROR").withError(
-          Number.parseInt(error.code || "500"),
-          "Error when comunicating with the back-end."
-        );
-      } else {
-        console.error("Unexpected error:", error);
-
-        return new ResponseModel("APP-ERROR").withError(
-          500,
-          "Error when trying to get session info, please try again later."
-        );
-      }
-    });
-};
-
-
-export const queryPublicLanguages = (
-  filter: LanguagesFilter
-): Promise<ResponseModel<Language[]>> => {
-  return LANGUAGES_CLIENT.get(`/v2/languages/public`, {
     params: filter,
   })
     .then((response) => response.data)
