+++ src/Domain/ProductLineEngineering/Entities/Language.ts
@@ -8,7 +8,6 @@ export class Language {
   type: string;
   stateAccept: string;
   semantics: string;
-  accessLevel?: string;
   userId?: string;
   ownerName?: string;
 
@@ -20,7 +19,6 @@ export class Language {
     type: string,
     stateAccept?: string,
     semantics?: string,
-    accessLevel?: string,
     userId?: string,
     ownerName?: string
   ) {
@@ -32,7 +30,6 @@ export class Language {
     this.stateAccept = stateAccept;
     this.semantics = semantics;
     this.userId = userId;
-    this.accessLevel = accessLevel;
     this.ownerName = ownerName;
   }
 }
