+++ src/core/components/LanguageManager/index.tsx
@@ -1,12 +1,11 @@
 import { useSession } from "@variamosple/variamos-components";
 import { useEffect, useState } from "react";
 import { Col, Row, Tab, Tabs } from "react-bootstrap";
-import { LanguagesContainer } from "../LanguageTable/LanguagesContainer";
+import { PublicLanguagesContainer } from "../PublicLanguages/PublicLanguagesContainer";
+import { UserLanguagesContainer } from "../UserLanguages/UserLangugesContainer";
 import CreateLanguageButton from "./CreateLanguageButton/CreateLanguageButton";
 import LanguageManagerLayout from "./LanguageManagerLayout/LanguageManagerLayout";
 import { LanguageManagerProps } from "./index.types";
-import {Button} from "react-bootstrap";
-import NoBackEndModal, {NoBackEndModalDefaultProps,NoBackEndModalProps} from "../NoBackEndModal";
 
 export default function LanguageManager({
   setLanguage,
@@ -15,29 +14,13 @@ export default function LanguageManager({
 }: LanguageManagerProps) {
   const { user } = useSession();
   const [isGuestUser, setIsGuestUser] = useState(true);
-  const [isLanguageDirectorUser, setIsLanguageDirectorUser] = useState(false);
-  const [loadLanguages, setLoadLanguages] = useState(false);
-  const [loadUserLanguages, setLoadUserLanguages] = useState(true);
   const [loadPublicLanguages, setLoadPublicLanguages] = useState(false);
-
-  /*To be delete in the end */
-  const [noBackEndModalState, setNoBackEndModalState] = useState<NoBackEndModalProps>({...NoBackEndModalDefaultProps});
-    const NoBackEndPopUp = () => {
-      setNoBackEndModalState({
-        ...NoBackEndModalDefaultProps,
-        show: true,
-        onCancel: () => setNoBackEndModalState((currentState) => ({...currentState, show: false})),
-      });
-    }
-  /*--------------------------*/
 
   useEffect(() => {
     const isGuest = user.roles.find((role) => role.toLowerCase() === "guest");
-    const isLanguageDirector = user.roles.find((role) => role.toLowerCase() === "language director");
-    console.log("User roles: ",user.roles, "\nUser id : ", user.id)
+
     setIsGuestUser(!!isGuest);
-    setIsLanguageDirectorUser(!!isLanguageDirector);
-    setLoadLanguages(!!isGuest);
+    setLoadPublicLanguages(!!isGuest);
   }, [user]);
 
   const handleCreateClick = () => {
@@ -57,13 +40,10 @@ export default function LanguageManager({
         <Col as={Row}>
           <Col sm={6}>
             <CreateLanguageButton handleCreateClick={handleCreateClick} />
-            <button
-            >Sementic Rules</button>
           </Col>
         </Col>
 
-        <LanguagesContainer
-          variant = "public"
+        <PublicLanguagesContainer
           loadDataOnInit={loadPublicLanguages}
           onLanguageClick={handleClick}
         />
@@ -73,33 +53,20 @@ export default function LanguageManager({
 
   return (
     <LanguageManagerLayout>
-
-      <div className='d-flex gap-1'>
-        <CreateLanguageButton handleCreateClick={handleCreateClick} />
-        
-        <Button 
-          variant="secondary"
-          onClick={NoBackEndPopUp}>
-          Sementic Rules
-        </Button>
-      </div>
+      <Col as={Row}>
+        <Col sm={6}>
+          <CreateLanguageButton handleCreateClick={handleCreateClick} />
+        </Col>
+      </Col>
 
       <Tabs
         defaultActiveKey="userLanguages"
         id="uncontrolled-tab"
-        onSelect={
-          (eventKey)=> {
-            switch (eventKey) {
-              case ("userLanguages"):
-                setLoadUserLanguages(true);
-                break;
-              case "publicLanguages":
-                setLoadPublicLanguages(true);
-                break;
-              case "allLanguages":
-                setLoadLanguages(true);
-                break;}
-              }}
+        onSelect={(eventKey) => {
+          if (eventKey === "publicLanguages") {
+            setLoadPublicLanguages(true);
+          }
+        }}
       >
         <Tab
           eventKey="userLanguages"
@@ -107,11 +74,7 @@ export default function LanguageManager({
           className="pt-3"
           unmountOnExit
         >
-          <LanguagesContainer
-            variant = "user"
-            loadDataOnInit={loadUserLanguages}
-            onLanguageClick={handleClick}
-          />
+          <UserLanguagesContainer onLanguageClick={handleClick} />
         </Tab>
 
         <Tab
@@ -120,27 +83,12 @@ export default function LanguageManager({
           className="pt-3"
           unmountOnExit
         >
-          <LanguagesContainer
-            variant = "public"
+          <PublicLanguagesContainer
             loadDataOnInit={loadPublicLanguages}
             onLanguageClick={handleClick}
           />
         </Tab>
-        { isLanguageDirectorUser && (<Tab 
-          eventKey="allLanguages"
-          title="All Languages"
-          className="pt-3"
-          unmountOnExit
-          >
-            <LanguagesContainer
-            variant = "languageDirector"
-            loadDataOnInit={loadLanguages}
-            onLanguageClick={handleClick}
-            />
-        </Tab>)}
       </Tabs>
-    {/* To be deleted in the end */}
-    <NoBackEndModal {...noBackEndModalState} />
     </LanguageManagerLayout>
   );
 }
