+++ src/index.css
@@ -4,20 +4,6 @@ body,
   height: 100%;
   margin: 0;
   padding: 0;
-
-  /*Couleurs principales ajoutées pour les languages*/
-  --red-color : #ff0000;
-  --red-color-contrast : #990000;
-  --red-color-text: #ffffff;
-
-  --green-color : #00AB17;
-  --green-color-contrast : #006600;
-  --green-color-text: #ffffff;
-
-  --yellow-color : #FFC107;
-  --yellow-color-contrast : #996600;
-  --yellow-color-text: #ffffff;
-  /*---------------------------- */
 }
 
 body {
@@ -51,51 +37,3 @@ body {
   text-align: center;
   color: #c0c0c0 !important;
 }
-
-
-/*Style pour les boutons */
-.btn-Variamos-red {
-  color: var(--red-color-text);
-  background-color: var(--red-color);
-  border-color: var(--red-color);
-}
-
-.btn-Variamos-red:hover {
-  color: var(--red-color-text);
-  background-color: var(--red-color-contrast);
-  border-color: var(--red-color-contrast);
-}
-
-.btn-Variamos-green {
-  color: var(--green-color-text);
-  background-color: var(--green-color);
-  border-color: var(--green-color);
-}
-
-.btn-Variamos-green:hover {
-  color: var(--green-color-text);
-  background-color: var(--green-color-contrast);
-  border-color: var(--green-color-contrast);
-}
-
-.btn-Variamos-yellow {
-  color: var(--yellow-color-text);
-  background-color: var(--yellow-color);
-  border-color: var(--yellow-color);
-}
-
-.btn-Variamos-yellow:hover {
-  color: var(--yellow-color-text);
-  background-color: var(--yellow-color-contrast);
-  border-color: var(--yellow-color-contrast);
-}
-
-.center {
-  text-align: center;
-  margin: auto;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-/*----------------------------*/
